Require minimum button push time to activate.
I was seeing the garage door spontaniously open from random power
drops/spikes. Logs showed that the controller thought it's button was
pushed. In an attempt to fix this, only count the garage door button
as pushed if it's held down for at least 50ms.

Also, some method renaming/reorganizing.